pyaml back, h5py too.
